refactor(phase-3b): extract content-url helper out of the db layer
- Move getContentUrl (absolute URL builder, no DB access) from lib/db/file.ts
  into a new lib/content-url.ts; update the 6 v3 import sites (import path
  only — calls and absolute output unchanged)
- Add a contentPath(id) relative helper and replace the server-side inline
  /api/contents/${id} literals with it (output-identical at every site)

v3 absolute URLs byte-identical; client-side literal left untouched.
lint + tsc green.
