+++ src/app/[lang]/(store)/library/actions.ts
@@ -1,5 +1,6 @@
 import "server-only";
 import { getDbAsync } from "@/prisma";
+import { contentPath } from "@/lib/content-url";
 import { guessCurrency } from "@/lib/currency";
 
 type ListedPackage = {
@@ -81,7 +82,7 @@ export async function retrievePackages(
     tmp
       .map((up) => up.package)
       .map(async (pkg) => {
-        const url = pkg.iconFile && `/api/contents/${pkg.iconFile.id}`;
+        const url = pkg.iconFile && contentPath(pkg.iconFile.id);
 
         return {
           id: pkg.id,
