+++ src/app/api/v3/[[...route]]/user.ts
@@ -4,7 +4,7 @@ import { getDbAsync } from "@/prisma";
 import { apiErrorResponse } from "@/lib/api/error";
 import type { Prisma } from "@prisma/client";
 import { getUserId } from "@/lib/api/auth";
-import { getContentUrl } from "@/lib/db/file";
+import { getContentUrl } from "@/lib/content-url";
 
 export async function getUserProfile(query: Prisma.ProfileWhereInput) {
   const prisma = await getDbAsync();
