+++ src/lib/db/file.ts
@@ -1,14 +1,6 @@
 import "server-only";
 import { getDbAsync } from "@/prisma";
 import type { PrismaTransaction } from "./transaction";
-import { headers } from "next/headers";
-
-export async function getContentUrl(id?: string | null) {
-  if (!id) return null;
-  const url = (await headers()).get("x-url") as string;
-  const origin = new URL(url).origin;
-  return `${origin}/api/contents/${id}`;
-}
 
 export async function retrieveFilesByUserId({
   userId,
