+++ src/lib/store-utils.ts
@@ -1,5 +1,6 @@
 import "server-only";
 import { getDbAsync } from "@/prisma";
+import { contentPath } from "@/lib/content-url";
 import { guessCurrency } from "./currency";
 import { existsUserPaymentHistory } from "./db/user-payment-history";
 
@@ -103,14 +104,14 @@ export async function retrievePackage(name: string) {
     pkg.PackageScreenshot.map(async (item) => {
       return {
         ...item,
-        url: `/api/contents/${item.file.id}`,
+        url: contentPath(item.file.id),
       };
     }),
   );
 
   return {
     ...pkg,
-    iconFileUrl: pkg.iconFile && `/api/contents/${pkg.iconFile.id}`,
+    iconFileUrl: pkg.iconFile && contentPath(pkg.iconFile.id),
     PackageScreenshot: screenshots,
   };
 }
@@ -216,7 +217,7 @@ export async function retrievePackages(
 
     return await Promise.all(
       tmp.map(async (pkg) => {
-        const url = pkg.iconFileId && `/api/contents/${pkg.iconFileId}`;
+        const url = pkg.iconFileId && contentPath(pkg.iconFileId);
 
         return {
           id: pkg.id,
@@ -288,7 +289,7 @@ export async function retrievePackages(
 
   return await Promise.all(
     tmp.map(async (pkg) => {
-      const url = pkg.iconFileId && `/api/contents/${pkg.iconFileId}`;
+      const url = pkg.iconFileId && contentPath(pkg.iconFileId);
 
       return {
         id: pkg.id,
